+++ gadget_app/common.py
@@ -1,14 +1,14 @@
 # Misc common data and code
 #
 # T. Lloyd
-# 15 Sep 2025
+# 17 Sep 2025
 
 from micropython import const
 import asyncio
 import time
 
 # HAL priority levels
-HAL_PRIORITY_IDLE = const(0)
+HAL_PRIORITY_IDLE = const(1)
 HAL_PRIORITY_MENU = const(10)
 HAL_PRIORITY_SHUTDOWN = const(100)